requirements.txt
# Core ML & data
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ESM-2 protein language model (Meta AI)
fair-esm>=2.0.0

# Bioinformatics
biopython>=1.81

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Dimensionality reduction (for UMAP capsid space visualization)
umap-learn>=0.5.3

# Interpretability / gradient attribution
captum>=0.6.0

# Notebook
jupyter>=1.0.0
ipython>=8.0.0

# Pipeline utilities
requests>=2.28.0
tqdm>=4.65.0
